Install default signal handlers
Default signal handlers are now installed at process-creation time.
With the exception of ignored signals, this is implemented using actual
signal handling functions, define in sighandlers.c.  Additionally, the
constant _TELOS_SIGMAX is defined in <signal.h>, which gives the total
number of signals available in Telos.

+++ src/dispatch/signal.c
@@ -165,9 +165,9 @@ void sys_signal (int sig, void(*func)(int)) {
 
     current->rc = (int) current->sigactions[sig].sa_handler;
 
-    if ((int) func == SIG_IGN) {
+    if (func == SIG_IGN) {
        current->sig_accept &= ~(1 << sig);
-    } else if ((int) func == SIG_DFL) {
+    } else if (func == SIG_DFL) {
         // ???
     } else {
         current->sigactions[sig].sa_handler = func;
